# Optimize restore_savepoint()