--- a/src/tree_store/mod.rs
+++ b/src/tree_store/mod.rs
@@ -17,8 +17,8 @@ pub(crate) use btree_base::{
 pub(crate) use btree_iters::{AllPageNumbersBtreeIter, BtreeExtractIf, BtreeRangeIter};
 pub use page_store::{file_backend, InMemoryBackend, Savepoint};
 pub(crate) use page_store::{
-    Page, PageHint, PageNumber, SerializedSavepoint, TransactionalMemory, FILE_FORMAT_VERSION2,
-    MAX_PAIR_LENGTH, MAX_VALUE_LENGTH, PAGE_SIZE,
+    BuddyAllocator, Page, PageHint, PageNumber, SerializedSavepoint, TransactionalMemory,
+    FILE_FORMAT_VERSION2, MAX_PAIR_LENGTH, MAX_VALUE_LENGTH, PAGE_SIZE,
 };
 pub(crate) use table_tree::{FreedPageList, FreedTableKey, TableTree, TableTreeMut};
 pub(crate) use table_tree_base::{InternalTableDefinition, TableType};